requirements.txt
pyyaml==5.1
Flask==1.0.2
Flask-Cors==3.0.3
gunicorn==19.7.1
pylint==1.8.2
pymongo==3.3.1
pytest==3.0.6
requests==2.22.0
flask-jwt-extended==3.21.0
flask-mail==0.9.1
flask-bcrypt==0.7.1
